requirements.txt
anyio==3.6.2
cffi==1.15.1
click==8.1.3
cryptography==38.0.4
fastapi==0.88.0
h11==0.14.0
idna==3.4
peewee==3.15.4
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
sniffio==1.3.0
starlette==0.22.0
typing==3.7.4.3
typing_extensions==4.4.0
uvicorn==0.20.0
